Restructured Server Node to Increase Stability of Subsequent Action Calls
The server worked well for the first action call.  However, after that action call ended (and especially if a second was made), the interactive marker controls would begin to behave erratically.  They would blink in and out, and it was difficult to manipulate them.

Playing with the code suggested that this behavior was due to the creation of new objects (listeners, broadcasters, utils, etc.) newly created for each action call.  They were trying to open already-open topics, orphaning old ones, and generally getting tangled up with the previous instantiations.

This commit changes many objects from being created for each action call to being created on node initialization and updated for each action call.  This seems to improve behavior for subsequent action calls.